fix: resolve issues #254, #255, #256, #257 - compilation errors and type conflicts
Issue #254: Fix NebulaLayout/Rarity type conflicts between nebula_explorer
           and nebula_gen modules by adding proper type imports.

Issue #255: Complete harvest type migration - fix remaining old API
           references, add contracterror/contracttype attributes to
           HarvestError/HarvestResult, implement AssetId and
           ResourceKey types, update gifting and ship_upgrade modules.

Issue #256: Fix missing imports and type resolution across the codebase:
           - Add missing String/Address imports in metadata_resolver
             and energy_manager
           - Fix soroban_sdk::i128 import error
           - Fix symbol_short bad character in events.rs
           - Fix owner not found in ship_nft::damage_ship
           - Fix leaderboard type mismatches
           - Fix access_control usize/u32 comparison

Issue #257: Fix borrow checker, Vec<u8>, and Soroban SDK compatibility:
           - Fix storage_uri and namespace moved value errors
           - Replace Vec<u8> with Bytes in cache_ttl_manager
           - Replace Symbol::to_string() calls with symbol_short or
             Symbol::new comparisons
           - Fix &Address reference handling in events
           - Fix Topics trait satisfaction for event publishing
           - Fix log! macro env reference issues in nebula_gen
           - Fix NebulaGen::has_anomaly env parameter type

Resolves #254, #255, #256, #257

+++ src/access_control.rs
@@ -478,7 +478,7 @@ pub fn grant_role_batch(
     expiry: Option<u32>,
 ) -> Result<(), AccessControlError> {
     // Check batch size FIRST before any other operation
-    if grantees.len() > BATCH_GRANT_LIMIT {
+    if grantees.len() as usize > BATCH_GRANT_LIMIT {
         return Err(AccessControlError::BatchLimitExceeded);
     }
